demo-code/requirements.txt
apiritif==0.6.5
Appium-Python-Client==0.26
astunparse==1.5.0
atomicwrites==1.1.5
attrs==18.1.0
bzt==1.11.1
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.9
colorlog==3.1.4
cssselect==1.0.3
decorator==4.3.0
EasyProcess==0.2.3
Flask==1.0.2
funcsigs==1.0.2
gevent==1.3.1
greenlet==0.4.13
hdrpy==0.3.1
idna==2.6
ipaddress==1.0.22
itsdangerous==0.24
Jinja2==2.10
jsonpath-rw==1.4.0
locustio==0.8.1
lxml==4.2.1
MarkupSafe==1.0
more-itertools==4.2.0
msgpack-python==0.5.6
nose==1.3.7
numpy==1.14.3
pluggy==0.6.0
ply==3.11
progressbar33==2.4
psutil==5.4.5
py==1.5.3
pytest==3.6.0
PyVirtualDisplay==0.2.1
PyYAML==3.12
pyzmq==17.0.0
requests==2.18.4
selenium==3.12.0
six==1.11.0
urllib3==1.22
urwid==2.0.1
Werkzeug==0.14.1
